requirements-dev.txt
-r requirements.txt
pytest>=7.0
pytest-django>=4.5
black>=23.0
flake8>=6.0
mypy>=1.0
beartype>=0.18.5
cachetools>=5.3.3
diskcache>=5.6.3
pathvalidate>=3.2.0
opentelemetry-sdk>=1.28.0
opentelemetry-exporter-otlp>=1.28.0
opentelemetry-instrumentation-django>=0.49b0

